Do not depend directly on bitcoin_hashes
`rust-bitcoin` exports the `bitcoin_hashes` dependency as
`bitcoin::hashes`. Using it instead of depending directly on
`bitcoin_hashes` reduces the maintenance burden when bitcoin upgrades
it's `bitcoin_hashes` dependency.

`bitcoin v0.29.2` depends on `bitcoin_hashes v0.11` so this change does
not effect the `bitcoin_hashes` version in use.

Use the public re-export of `bitcoin_hashes` from `bitcoin::hashes`.

This is not an API breaking change.