Do not depend directly on bech32
`rust-bitcoin` exports the `rust-bech32` dependency as
`bitcoin::bech32`. Using it instead of depending directly on
`rust-bech32` reduces the maintenance burden when bitcoin upgrades it's
`rust-bech32` dependency.

`bitcoin v0.29.2` depends on `bech32 v0.9` so this change does not
effect the `rust-bech32` version in use.

Use the public re-export of `bech32` from `bitcoin::bech32`.

Note we still depend on `bech32` in the `lightning-invoice` fuzz crate.

This is not an API breaking change.