Do not depend directly on secp256k1
`rust-bitcoin` exports the `rust-secp256k1` dependency as
`bitcoin::secp256k1`. Using it instead of depending directly on
`rust-secp256k1` reduces the maintenance burden when bitcoin upgrades
it's `rust-secp256k1` dependency.

`bitcoin v0.29.2` depends on `secp256k1 v0.24` so this change does not
effect the `rust-secp256k1` version in use. It does change the feature
set of the dependency.

Use the public re-export of `secp256k1` from `bitcoin::secp256k1`.

This is not an API breaking change.