Make channel_type optional in ChannelOpenRequest event.
Peers may send messages requesting to open a channel without specifying
the channel type. This gets inferred by LDK when the channel actually
opens, but can be passed to the event handler as an Option so the node
may factor the lack of specificity in to the channel opening decision.

This should fix #2804.