Unlock dynamic dispatch for SignerProvider.
This change makes `lightning::sign::WriteableEcdsaChannelSigner` [object safe](https://doc.rust-lang.org/reference/items/traits.html#object-safety),
so a SignerProvider can be built and assigned to a node dynamically based on a config entry. As a result it allows a cleaner separation of code
for different signers because now different SignerProvider implementations can live in different structs implementing the trait.