Clear pending flags when peer disconnects
Forget pending signature requests that will be recomputed when we receive a
`channel_reestablish`. This ensures that we don't e.g. send a `revoke_and_ack`
and `commitment_update` twice: once when signing is unblocked, and again when
receiving the counterparty's reestablish message. It also ensures that we
don't send a stray `channel_reestablish` if we are recconected, and then
disconnected again before making any progress.